+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added ValidatorProviderInterface for a class to act as a validator provider.
 - Added ConstraintValidator to validate if iterable values are allowed by array with constraints.
 - Added property to LengthValidator to disallow new line characters.
+- Added UrlValidator to validate URL string.
 
 ### Changed